backend/requirements.txt
asgiref==3.7.2
boto3==1.28.45
botocore==1.31.45
certifi==2023.7.22
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-storages==1.14
djangorestframework==3.14.0
drf-base64==2.0
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.6.2
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
